requirements.txt
transformers
sentence-transformers
torch
gradio
PyMuPDF
faiss-cpu